+++ src/1_Page/OAuthRedirect/index.js
@@ -1,32 +1,11 @@
-import { useNavigate } from "react-router-dom";
 import React from "react";
-import { useCookies } from "react-cookie";
-import useLoginWithKakao from "../../3_Entity/Account/useLoginWithKakao";
-import useGetMyInfo from "../../3_Entity/Account/useGetMyInfo";
+import useLogin from "../../3_Entity/Account/useLogin";
 const OAuthRedirect = () => {
-  const navigate = useNavigate();
-  const [cookies, setCookies] = useCookies([
-    "accessToken",
-    "refreshToken",
-    "userIdx",
-  ]);
   const queryParams = new URLSearchParams(window.location.search);
-  const [kakaoLoading] = useLoginWithKakao(
+  const [serverState] = useLogin(
     queryParams.get("code"),
     queryParams.get("state")
   );
-  console.log(queryParams.get("code"), "그리고",
-  queryParams.get("state"))
-  // const [userinfo] = useGetMyInfo(kakaoLoading);
-
-  // React.useEffect(() => {
-  //   const expires = new Date();
-  //   expires.setMinutes(expires.getMinutes() + 30);
-  //   if (!kakaoLoading && userinfo) {
-  //     setCookies("userIdx", userinfo.idx, { path: "/", expires });
-  //     navigate("/");
-  //   }
-  // }, [kakaoLoading, userinfo]);
 
   return <div>Authenticating...</div>;
 };
