+++ src/1_Page/index.js
@@ -20,7 +20,6 @@ const Page = () => {
         <Route path="/sns" element={<Sns />} />
         <Route path="/tracking" element={<Tracking />} />
         <Route path="/auth/callback" element={<OAuthRedirect />} />
-        {/* <Route path="*" element={} /> 에러 페이지 추가 예정 */}
       </Routes>
     </STYLE.Main>
   );
